STYLE: Refine ClickerViewController
This commit refines some comments in ClickerViewController for better
readability and evaluation of code.

ClickerViewController times the user how fast they can tap on the screen
and returns the value back to ScoresViewController (still yet to be
created).

In future commits, amendments to the code in ClickerViewController may
be changed, whenever the developers see fit.