FEAT: Add ScoresTableViewController to project
In this commit, ScoresTableViewController has been added into the
project and is working normally.

ScoresTableViewController consists of an unwind segue which retrieves
the timings as a score for the user, and then presents it to the user in the
form of a Table View.

In future commits, changes to the code may be made as a whole or partially
to the project, wherever the developers see fit.