FEAT: Added swipe-down gesture
In this commit, a swipe-down gesture has been added into
ClickerViewController, which cancels the clicker game. In the event that
the game has started, an alert comes up intervening the timer and the
cancel asking if the user wants to cancel. If true, then the timer is
invalidated and the View Controller is dismissed, and vice versa.

Besides the gesture, an application icon has also been added
into the application for better design.

In future commits, these gestures and the application icon may
change.